+++ api/src/main/java/ru/flamexander/transfer/service/core/api/dtos/ExecuteTransferDtoResponse.java
@@ -1,0 +1,13 @@
+package ru.flamexander.transfer.service.core.api.dtos;
+
+//import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@NoArgsConstructor
+@Data
+//@Schema(description = "ДТО для создания нового счета")
+public class ExecuteTransferDtoResponse {
+}
