+++ backend/src/main/java/ru/flamexander/transfer/service/core/backend/controllers/TransfersController.java
@@ -1,0 +1,17 @@
+package ru.flamexander.transfer.service.core.backend.controllers;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+import ru.flamexander.transfer.service.core.backend.services.TransferService;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/transfers")
+public class TransfersController {
+    private final TransferService transferService;
+
+    @PostMapping("/execute")
+    public void executeTransfer() {
+        transferService.transfer(2L, 3L);
+    }
+}
