+++ backend/src/main/java/ru/flamexander/transfer/service/core/backend/entities/Account.java
@@ -1,0 +1,43 @@
+package ru.flamexander.transfer.service.core.backend.entities;
+
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Entity
+@Table(name = "accounts")
+public class Account {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "client_id")
+    private Long clientId;
+
+    @Column(name = "account_number")
+    private String accountNumber;
+
+    @Column(name = "balance")
+    private BigDecimal balance;
+
+    @Column(name = "created_at")
+    @CreationTimestamp
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    @UpdateTimestamp
+    private LocalDateTime updatedAt;
+
+    public Account(Long clientId, BigDecimal balance) {
+        this.clientId = clientId;
+        this.balance = balance;
+    }
+}
