+++ backend/src/main/java/ru/flamexander/transfer/service/core/backend/errors/AppLogicException.java
@@ -1,0 +1,14 @@
+package ru.flamexander.transfer.service.core.backend.errors;
+
+public class AppLogicException extends RuntimeException {
+    private String code;
+
+    public String getCode() {
+        return code;
+    }
+
+    public AppLogicException(String code, String message) {
+        super(message);
+        this.code = code;
+    }
+}
