+++ backend/src/main/java/ru/flamexander/transfer/service/core/backend/errors/FieldValidationError.java
@@ -1,0 +1,11 @@
+package ru.flamexander.transfer.service.core.backend.errors;
+
+public class FieldValidationError {
+    private String fieldName;
+    private String message;
+
+    public FieldValidationError(String fieldName, String message) {
+        this.fieldName = fieldName;
+        this.message = message;
+    }
+}
