+++ backend/src/main/java/ru/flamexander/transfer/service/core/backend/errors/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package ru.flamexander.transfer.service.core.backend.errors;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
