--- src/main/java/ru/flamexander/java/pro/spring/dtos/AccountsPageDto.java
@@ -1,14 +1,0 @@
-package ru.flamexander.java.pro.spring.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class AccountsPageDto {
-    private List<AccountDto> items;
-}
