--- src/main/java/ru/flamexander/java/pro/spring/dtos/CreateAccountDto.java
@@ -1,10 +1,0 @@
-package ru.flamexander.java.pro.spring.dtos;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Data
-public class CreateAccountDto {
-    private Long initialBalance;
-}
