--- src/main/java/ru/flamexander/java/pro/spring/entities/Account.java
@@ -1,39 +1,0 @@
-package ru.flamexander.java.pro.spring.entities;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Entity
-@Table(name = "accounts")
-public class Account {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "client_id")
-    private Long clientId;
-
-    @Column(name = "balance")
-    private Long balance;
-
-    @Column(name = "created_at")
-    @CreationTimestamp
-    private LocalDateTime createdAt;
-
-    @Column(name = "updated_at")
-    @UpdateTimestamp
-    private LocalDateTime updatedAt;
-
-    public Account(Long clientId, Long balance) {
-        this.clientId = clientId;
-        this.balance = balance;
-    }
-}
