--- src/main/java/ru/flamexander/java/pro/spring/errors/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.flamexander.java.pro.spring.errors;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
